refactor: Moved EF generated models in Model class library
