+++ src/ServiceInsights.sln
@@ -33,6 +33,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CreateEpisodeTests", "..\te
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CreateUpdateEpisodeTests", "..\tests\EpisodeManagementServiceTests\CreateUpdateEpisodeTests\CreateUpdateEpisodeTests.csproj", "{3EDF6D3C-73A8-4C66-A6E4-83A4EE7E1F6A}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Model", "Shared\Model\Model.csproj", "{C232F5A4-F99A-4185-BE03-A490F8AB029B}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -82,6 +84,10 @@ Global
 		{3EDF6D3C-73A8-4C66-A6E4-83A4EE7E1F6A}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{3EDF6D3C-73A8-4C66-A6E4-83A4EE7E1F6A}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{3EDF6D3C-73A8-4C66-A6E4-83A4EE7E1F6A}.Release|Any CPU.Build.0 = Release|Any CPU
+		{C232F5A4-F99A-4185-BE03-A490F8AB029B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{C232F5A4-F99A-4185-BE03-A490F8AB029B}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{C232F5A4-F99A-4185-BE03-A490F8AB029B}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{C232F5A4-F99A-4185-BE03-A490F8AB029B}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
 		{EFF39066-513A-4054-8347-2E43958476A8} = {8BBD7332-93CD-4C28-878C-314BB44A6224}
@@ -91,5 +97,6 @@ Global
 		{E2797CCA-CD3C-4BF1-87CB-4D34FE8289A2} = {DACA4EB5-14F5-42D8-AD61-DA1FA4DBB6ED}
 		{853338EB-9D25-4B5D-895B-FBEC7870A68A} = {C1742A07-A12D-4894-849E-87DB9CEA03C3}
 		{69407060-E86D-4761-88B6-BCBAB09A481A} = {C1742A07-A12D-4894-849E-87DB9CEA03C3}
+		{C232F5A4-F99A-4185-BE03-A490F8AB029B} = {C1742A07-A12D-4894-849E-87DB9CEA03C3}
 	EndGlobalSection
 EndGlobal
