--- src/Shared/Data/Models/Episode.cs
@@ -1,9 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-
-namespace Data.Models;
-
-public partial class Episode
-{
-    public string EpisodeId { get; set; } = null!;
-}
