+++ src/Shared/Model/Episode.cs
@@ -1,0 +1,9 @@
+﻿using System;
+using System.Collections.Generic;
+
+namespace NHS.ServiceInsights.Model;
+
+public partial class Episode
+{
+    public string EpisodeId { get; set; } = null!;
+}
